Helpdesk additions for insertion report UNDONE
