+++ Plugins/Helpdesk/Bootstrap.php
@@ -6,8 +6,8 @@ use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use FrontendUserManagement\Services\AccountNavigationService;
 use Helpdesk\Controller\Backend\BackendHelpdeskController;
-use Helpdesk\Controller\Backend\BackendHelpdeskMessengerController;
 use Helpdesk\Controller\Frontend\FrontendHelpdeskController;
+use Helpdesk\Models\IssueTypeGroup;
 use Helpdesk\Models\IssueTypes;
 use Helpdesk\Models\Ticket;
 use Helpdesk\Services\HelpdeskMessengerService;
@@ -20,6 +20,7 @@ use Oforge\Engine\Modules\Core\Exceptions\ConfigElementAlreadyExistException;
 use Oforge\Engine\Modules\Core\Exceptions\ConfigOptionKeyNotExistException;
 use Oforge\Engine\Modules\Core\Exceptions\ParentNotFoundException;
 use Oforge\Engine\Modules\Core\Exceptions\ServiceNotFoundException;
+use Oforge\Engine\Modules\CRUD\Services\GenericCrudService;
 
 /**
  * Class Bootstrap
@@ -34,7 +35,6 @@ class Bootstrap extends AbstractBootstrap {
     public function __construct() {
         $this->endpoints = [
             BackendHelpdeskController::class,
-            BackendHelpdeskMessengerController::class,
             FrontendHelpdeskController::class,
         ];
 
@@ -46,6 +46,7 @@ class Bootstrap extends AbstractBootstrap {
         $this->models = [
             Ticket::class,
             IssueTypes::class,
+            IssueTypeGroup::class,
         ];
 
         $this->dependencies = [
@@ -60,10 +61,15 @@ class Bootstrap extends AbstractBootstrap {
      * @throws ServiceNotFoundException
      */
     public function install() {
+        /** @var GenericCrudService $crud */
+        $crud = Oforge()->Services()->get('crud');
+        $supportGroup = $crud->create(IssueTypeGroup::class, ['issueTypeGroupName' => 'support']);
+        $reportGroup = $crud->create(IssueTypeGroup::class, ['issueTypeGroupName' => 'report']);
+
         /** @var HelpdeskTicketService $helpdeskTicketService */
         $helpdeskTicketService = Oforge()->Services()->get('helpdesk.ticket');
-        $helpdeskTicketService->createIssueType('99 Problems');
-        $helpdeskTicketService->createIssueType('but the horse ain\'t one');
+        $helpdeskTicketService->createIssueType('99 Problems', $supportGroup);
+        $helpdeskTicketService->createIssueType('but the horse ain\'t one', $supportGroup);
 
         $helpdeskTicketService->createNewTicket(1, 1, 'but the horse ain\'t one',
             'Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.');
