+++ Plugins/Helpdesk/Controller/Backend/BackendHelpdeskController.php
@@ -7,6 +7,7 @@ use Doctrine\ORM\ORMException;
 use FrontendUserManagement\Models\User;
 use FrontendUserManagement\Services\UserService;
 use Helpdesk\Models\Ticket;
+use Helpdesk\Services\HelpdeskMessengerService;
 use Helpdesk\Services\HelpdeskTicketService;
 use Oforge\Engine\Modules\AdminBackend\Core\Abstracts\SecureBackendController;
 use Oforge\Engine\Modules\Auth\Models\User\BackendUser;
@@ -55,13 +56,14 @@ class BackendHelpdeskController extends SecureBackendController {
 
     /**
      * @param Request $request
-     * @param Response $response
+     * @param Response $response*
+     * @EndpointAction(path="/closeTicket")
      *
      * @return Response
      * @throws ORMException
      * @throws OptimisticLockException
      * @throws ServiceNotFoundException
-     * @EndpointAction()
+     *
      */
     public function closeTicketAction(Request $request, Response $response) {
         if ($request->isPost()) {
@@ -73,6 +75,76 @@ class BackendHelpdeskController extends SecureBackendController {
             $helpdeskTicketService->changeStatus($ticketId, 'closed');
 
             return $response->withRedirect('/backend/helpdesk');
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @param Response $response
+     * @EndpointAction(path="/messenger/{id}")
+     *
+     * @throws ORMException
+     * @throws OptimisticLockException
+     * @throws ServiceNotFoundException
+     */
+    public function messengerAction(Request $request, Response $response, array $args) {
+        /** @var Router $router */
+        $router = Oforge()->App()->getContainer()->get('router');
+        $uri = $router->pathFor('backend_helpdesk');
+
+        if (isset($args) && isset($args['id'])) {
+            $ticketId = $args['id'];
+
+            if ($request->isPost()) {
+                $targetId = $args['id'];
+
+                $senderId = $request->getParsedBody()['sender'];
+                $message  = $request->getParsedBody()['message'];
+
+                if (!$senderId) {
+                    $senderId = 'helpdesk';
+                }
+
+                /** @var HelpdeskMessengerService $helpdeskMessengerService */
+                $helpdeskMessengerService = Oforge()->Services()->get('helpdesk.messenger');
+
+                $conversation = $helpdeskMessengerService->getConversationsByTarget($targetId, $senderId);
+
+                if (sizeof($conversation) > 0) {
+                    $conversation = $conversation[0];
+                }
+
+                $helpdeskMessengerService->sendMessage($conversation['id'], 'helpdesk', $senderId, $message);
+
+                $uri = $router->pathFor('backend_helpdesk_messenger');
+                return $response->withRedirect($uri . '/' . $args['id'], 302);
+            }
+
+            /** @var HelpdeskTicketService $helpdeskTicketService */
+            $helpdeskTicketService = Oforge()->Services()->get('helpdesk.ticket');
+
+            $ticketData = $helpdeskTicketService->getTicketById($ticketId);
+
+            /** @var HelpdeskMessengerService $helpdeskMessengerService */
+            $helpdeskMessengerService = Oforge()->Services()->get('helpdesk.messenger');
+
+            $conversation = $helpdeskMessengerService->getConversationsByTarget($ticketId, 'helpdesk');
+
+            if (sizeof($conversation) > 0) {
+                $conversation = $conversation[0];
+            }
+
+            $messages = $helpdeskMessengerService->getMessagesOfConversation($conversation['id']);
+            if (!isset($messages[0])) {
+                return $response->withRedirect($uri, 302);
+            }
+
+            Oforge()->View()->assign([
+                'messages' => $messages,
+                'ticket'   => $ticketData,
+            ]);
+        } else {
+            return $response->withRedirect($uri, 302);
         }
     }
